Resolve #3
ItemHandlers of the machines were working on the copies of the itemstacks, when minecraft slot-click logic requires a direct itemstack reference to modify the slot values, which caused troubles.

I'm fairly sure this shouldn't break anything, but worth noting this down if I'm wrong as this affects all machines apparently.