Release 1.0.1 (#4)
* Resolve #3

ItemHandlers of the machines were working on the copies of the itemstacks, when minecraft slot-click logic requires a direct itemstack reference to modify the slot values, which caused troubles.

I'm fairly sure this shouldn't break anything, but worth noting this down if I'm wrong as this affects all machines apparently.

* Resolve #1

* Additonal changes to the XU2 Gradle configs

Disabled Asset related tasks to speed up the boot preparation times - those tasks are broken and can take up to like 20 minutes to complete without cutting network access to force timeouts on all connections.

Added Railcraft as runtime dependency in-dev

Made the game wait for the debugger on launch.

TODO: Fix the console log output.

* Update README.md